pyproject.toml
[project]
name = "kafka-exp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiokafka>=0.10.0",
    "confluent-kafka>=2.12.2",
    "httpx>=0.25.0",
    "kafka-python==2.0.2",
    "langflow>=1.0.0",
]
